Initialization auth fix
* There were still some problems in the initialization of apps. If a
user somehow opened gui at the /api-callback page, it would not force
him to log in.
* The new implementation should detect such situations and redirect the
user to the oidc server to log in.
* Also, the Server down error should be now at least a bit smarter. It
should show the server down error if the response error status is '0'
which is returned if the server is down. Otherwise, it should show an
error message in the error that was thrown.